backend/requirements.txt
fastapi
uvicorn
python-docx
mammoth
openai
python-multipart